Enemy AI updates to movement algorithm

+++ Rock_Paper_Scissors/Assets/Scripts/Enemy/EnemyAIAction.cs
@@ -6,5 +6,5 @@ public class EnemyAIAction
 {
     public GridObject gridObject;
     public UnitAction unitAction;
-    public int actionValue;
+    public float actionValue;
 }
